+++ client/src/App.tsx
@@ -1,14 +1,16 @@
 import { Outlet } from "react-router-dom"
 
-import Nav from "./components/Nav Folder/Nav";
+import Nav from "./components/Nav Component Folder/Nav";
+import Footer from "./components/Footer Component Folder/Footer";
 
 function App() {
   return (
     <>
-    <Nav />
-      <main>
-        <Outlet />
-      </main>
+      <Nav />
+        <main>
+          <Outlet />
+        </main>
+      <Footer />
     </>
   );
 }
