+++ client/src/components/Dashboard.tsx
@@ -1,3 +1,9 @@
+// React Hook Imports
+
+// MUI Material Imports
+
+// CSS Imports
+
 const DashboardPage = () => {
 
 }
