+++ client/src/components/Nav Component Folder/nav.module.css
@@ -1,6 +1,6 @@
-.button:hover {
+/* .button:hover {
     color: red;
-}
+} */
 
 .currency {
     background-color: hsla(108, 100%, 50%, 0.5);
