+++ client/src/components/charts.tsx
@@ -1,4 +1,12 @@
+// React Hook Imports
+
+// MUI Material Imports
+
+// MUI Chart Imports
 import { LineChart } from '@mui/x-charts/LineChart';
+
+// CSS Imports
+
 
 export default function ChartsOverviewDemo() {
   return (
