+++ client/src/pages/Error Page Folder/ErrorPage.tsx
@@ -1,6 +1,13 @@
-import classes from '../Error Page Folder/error.module.css';
+// React Hook Imports
+
+// React Component Imports
+
+// MUI Material Imports
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
+
+// CSS Imports
+import classes from '../Error Page Folder/error.module.css';
 
 const ErrorPage = () => {
     return (
