+++ client/src/pages/Landing Page Folder/LandingPage.tsx
@@ -1,5 +1,12 @@
-import LoginCard from "../../components/Login Folder/LoginCard";
-import SignUpCard from "../../components/Sign Up Folder/SignUp";
+// React Hook Imports
+
+// React Component Imports
+import LoginCard from "../../components/Login Component Folder/LoginCard";
+import SignUpCard from "../../components/Sign Up Component Folder/SignUp";
+
+// MUI Material Imports
+
+// CSS Imports
 
 const LandingPage = () => {
     return (
