+++ client/src/pages/Main Page Folder/MainPage.tsx
@@ -1,6 +1,13 @@
-import StockCard from "../../components/Stock Card Folder/StockCard";
-import SearchBar from "../../components/Search Folder/SearchBar";
+// React Hook Imports
+
+// React Component Imports
+import StockCard from "../../components/Stock Card Component Folder/StockCard";
+import SearchBar from "../../components/Search Component Folder/SearchBar";
+
+// MUI Material Imports
 import Box from "@mui/material/Box";
+
+// CSS Imports
 
 const MainPage = () => {
